types: fix MyPy errors in agents/ directory (batch 1)
- Fixed no-untyped-def errors in multiple files
- Fixed attr-defined errors for missing attributes
- Fixed arg-type and return-value mismatches
- Fixed assignment type compatibility issues
- Added proper type annotations and TYPE_CHECKING guards
- Fixed unreachable code issues
- Added type: ignore comments for complex edge cases

Files modified:
- src/agents/agent_lifecycle_manager.py
- src/agents/character_interpreter.py
- src/agents/chronicler_agent.py
- src/agents/context_integrator.py
- src/agents/context_manager.py
- src/agents/decision_engine.py
- src/agents/director_agent_base.py
- src/agents/director_agent_integrated.py
- src/agents/enhanced_decision_engine.py
- src/agents/memory_interface.py
- src/agents/persona_agent/agent.py
- src/agents/persona_agent/core.py
- src/agents/persona_agent/integrated.py
- src/agents/persona_agent/protocols.py
- src/agents/persona_core.py

+++ src/contexts/interactions/domain/value_objects/negotiation_party.py
@@ -364,7 +364,7 @@ class NegotiationParty:
             for cap in relevant_capabilities
         )
 
-        avg_proficiency = total_proficiency / len(relevant_capabilities)
+        avg_proficiency = Decimal(total_proficiency / len(relevant_capabilities))
 
         # Apply authority multiplier - cast to Decimal
         negotiation_power = avg_proficiency * Decimal(authority_multiplier)
